Refactor invitation utilities into dedicated module
